feat(http): add upload client ops and sync backend openapi to a1bcdf3 (#149)
## Summary

- Re-vendor the backend OpenAPI spec (`api/specs/backend/v04.yaml`) from
openapi `main` (a1bcdf3) and regenerate the `libs/backend-api` Rust
models: presigned-upload schemas (`PresignedUpload`,
`UploadRequiredHeaders`) removed, upload-credential schemas
(`UploadWithCredentials`, `UploadCredentials`, S3/GCS variants) added,
`stability_window_secs` widened to int64, and upload source timestamp
fields renamed.
- Absorb the model fallout in agent source and tests: carry
`stability_window_secs` as `i64` end-to-end through the domain model and
scan/disk call sites (no truncating casts), and drop removed fields from
fixtures.
- Add HTTP client operations `uploads::{create, vend_credentials,
confirm}` in `agent/src/http/uploads.rs` following the existing
`devices.rs` pattern, with `MockClient` routing and tests asserting
method, path, body, and auth token.

Runtime wiring of the upload ops into `services::backend` is
intentionally out of scope (see plan decision log:
`plans/active/20260713-sync-openapi-spec-and-upload-client-ops.md`).


<!-- codesmith:footer -->
---
<a
href="https://app.blacksmith.sh/mirurobotics/codesmith/agent/pr/149"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-dark-v2.svg"><source
media="(prefers-color-scheme: light)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-light-v2.svg"><img
alt="View with Codesmith"
src="https://pr-comments-assets.blacksmith.sh/codesmith/view-with-codesmith-dark-v2.svg"></picture></a>
<a
href="https://backend.blacksmith.sh/track/enable-autofix?expires=1786588470&installation_id=142588636&pr_number=149&repository=mirurobotics%2Fagent&return_to=https%3A%2F%2Fgithub.com%2Fmirurobotics%2Fagent%2Fpull%2F149&signature=6eaf048962066aa5922bbcc3f6b7a52dc29830f55f63bc668d370009e8232923"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-dark.svg"><source
media="(prefers-color-scheme: light)"
srcset="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-light.svg"><img
alt="Autofix with Codesmith"
src="https://pr-comments-assets.blacksmith.sh/codesmith/autofix-with-codesmith-dark.svg"></picture></a>
<sup>Need help on this PR? Tag <code>/codesmith</code> with what you
need. Autofix is disabled.</sup>

<!-- codesmith:autofix:disabled -->
<!-- /codesmith:footer -->

---------

+++ agent/src/disk/upload_rules.rs
@@ -50,7 +50,7 @@ mod tests {
     // =============================== TEST HELPERS ================================= //
 
     /// Build an UploadRule from Default with only the source fields set.
-    fn rule_with(id: &str, glob: &str, stability_window_secs: i32) -> UploadRule {
+    fn rule_with(id: &str, glob: &str, stability_window_secs: i64) -> UploadRule {
         UploadRule {
             id: id.to_string(),
             source: UploadRuleSource {
