+++ agent/tests/workers/scan_bridge.rs
@@ -131,7 +131,7 @@ impl Stores {
 }
 
 /// Build an UploadRule with only its id + source fields set.
-fn rule_with(id: &str, glob: &str, stability_window_secs: i32) -> UploadRule {
+fn rule_with(id: &str, glob: &str, stability_window_secs: i64) -> UploadRule {
     UploadRule {
         id: id.to_string(),
         source: UploadRuleSource {
